+++ pkg/instrumentors/bpf/database/sql/probe.go
@@ -12,14 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package server
+package sql
 
 import (
 	"bytes"
 	"encoding/binary"
 	"errors"
 	"os"
+	"strconv"
 
+	"go.opentelemetry.io/auto/pkg/inject"
 	"go.opentelemetry.io/auto/pkg/instrumentors/bpffs"
 
 	"github.com/cilium/ebpf"
@@ -27,7 +29,6 @@ import (
 	"github.com/cilium/ebpf/perf"
 	"golang.org/x/sys/unix"
 
-	"go.opentelemetry.io/auto/pkg/inject"
 	"go.opentelemetry.io/auto/pkg/instrumentors/context"
 	"go.opentelemetry.io/auto/pkg/instrumentors/events"
 	"go.opentelemetry.io/auto/pkg/instrumentors/utils"
@@ -39,17 +40,16 @@ import (
 
 //go:generate go run github.com/cilium/ebpf/cmd/bpf2go -target amd64,arm64 -cc clang -cflags $CFLAGS bpf ./bpf/probe.bpf.c
 
-const instrumentedPkg = "net/http"
+const instrumentedPkg = "database/sql"
 
-// Event represents an event in an HTTP server during an HTTP
+// Event represents an event in an SQL database
 // request-response.
 type Event struct {
 	context.BaseSpanProperties
-	Method [7]byte
-	Path   [100]byte
+	Query [100]byte
 }
 
-// Instrumentor is the net/http instrumentor.
+// Instrumentor is the database/sql instrumentor.
 type Instrumentor struct {
 	bpfObjects   *bpfObjects
 	uprobes      []link.Link
@@ -57,72 +57,78 @@ type Instrumentor struct {
 	eventsReader *perf.Reader
 }
 
+// IncludeDBStatementEnvVar is the environment variable to opt-in for sql query inclusion in the trace.
+const IncludeDBStatementEnvVar = "OTEL_GO_AUTO_INCLUDE_DB_STATEMENT"
+
 // New returns a new [Instrumentor].
 func New() *Instrumentor {
 	return &Instrumentor{}
 }
 
-// LibraryName returns the net/http package name.
+// LibraryName returns the database/sql/ package name.
 func (h *Instrumentor) LibraryName() string {
 	return instrumentedPkg
 }
 
-// FuncNames returns the function names from "net/http" that are instrumented.
+// FuncNames returns the function names from "database/sql" that are instrumented.
 func (h *Instrumentor) FuncNames() []string {
-	return []string{"net/http.HandlerFunc.ServeHTTP"}
+	return []string{"database/sql.(*DB).queryDC"}
 }
 
 // Load loads all instrumentation offsets.
 func (h *Instrumentor) Load(ctx *context.InstrumentorContext) error {
-	spec, err := ctx.Injector.Inject(loadBpf, "go", ctx.TargetDetails.GoVersion.Original(), []*inject.StructField{
+	spec, err := ctx.Injector.Inject(loadBpf, "go", ctx.TargetDetails.GoVersion.Original(), nil, []*inject.FlagField{
 		{
-			VarName:    "method_ptr_pos",
-			StructName: "net/http.Request",
-			Field:      "Method",
-		},
-		{
-			VarName:    "url_ptr_pos",
-			StructName: "net/http.Request",
-			Field:      "URL",
-		},
-		{
-			VarName:    "ctx_ptr_pos",
-			StructName: "net/http.Request",
-			Field:      "ctx",
-		},
-		{
-			VarName:    "path_ptr_pos",
-			StructName: "net/url.URL",
-			Field:      "Path",
-		},
-		{
-			VarName:    "ctx_ptr_pos",
-			StructName: "net/http.Request",
-			Field:      "ctx",
-		},
-		{
-			VarName:    "headers_ptr_pos",
-			StructName: "net/http.Request",
-			Field:      "Header",
-		},
-	}, nil, false)
+			VarName: "should_include_db_statement",
+			Value:   shouldIncludeDBStatement(),
+		}}, true)
 
 	if err != nil {
 		return err
 	}
 
 	h.bpfObjects = &bpfObjects{}
+
 	err = utils.LoadEBPFObjects(spec, h.bpfObjects, &ebpf.CollectionOptions{
 		Maps: ebpf.MapOptions{
 			PinPath: bpffs.PathForTargetApplication(ctx.TargetDetails),
 		},
 	})
+
 	if err != nil {
 		return err
 	}
 
-	for _, funcName := range h.FuncNames() {
-		h.registerProbes(ctx, funcName)
+	offset, err := ctx.TargetDetails.GetFunctionOffset(h.FuncNames()[0])
+
+	if err != nil {
+		return err
+	}
+
+	up, err := ctx.Executable.Uprobe("", h.bpfObjects.UprobeQueryDC, &link.UprobeOptions{
+		Address: offset,
+	})
+
+	if err != nil {
+		return err
+	}
+
+	h.uprobes = append(h.uprobes, up)
+
+	retOffsets, err := ctx.TargetDetails.GetFunctionReturns(h.FuncNames()[0])
+
+	if err != nil {
+		return err
+	}
+
+	for _, ret := range retOffsets {
+		retProbe, err := ctx.Executable.Uprobe("", h.bpfObjects.UprobeQueryDC_Returns, &link.UprobeOptions{
+			Address: ret,
+		})
+		if err != nil {
+			return err
+		}
+		h.returnProbs = append(h.returnProbs, retProbe)
 	}
 
 	rd, err := perf.NewReader(h.bpfObjects.Events, os.Getpagesize())
@@ -134,45 +140,9 @@ func (h *Instrumentor) Load(ctx *context.InstrumentorContext) error {
 	return nil
 }
 
-func (h *Instrumentor) registerProbes(ctx *context.InstrumentorContext, funcName string) {
-	logger := log.Logger.WithName("net/http-instrumentor").WithValues("function", funcName)
-	offset, err := ctx.TargetDetails.GetFunctionOffset(funcName)
-	if err != nil {
-		logger.Error(err, "could not find function start offset. Skipping")
-		return
-	}
-	retOffsets, err := ctx.TargetDetails.GetFunctionReturns(funcName)
-	if err != nil {
-		logger.Error(err, "could not find function end offsets. Skipping")
-		return
-	}
-
-	up, err := ctx.Executable.Uprobe("", h.bpfObjects.UprobeServerMuxServeHTTP, &link.UprobeOptions{
-		Address: offset,
-	})
-	if err != nil {
-		logger.V(1).Info("could not insert start uprobe. Skipping",
-			"error", err.Error())
-		return
-	}
-
-	h.uprobes = append(h.uprobes, up)
-
-	for _, ret := range retOffsets {
-		retProbe, err := ctx.Executable.Uprobe("", h.bpfObjects.UprobeServerMuxServeHTTP_Returns, &link.UprobeOptions{
-			Address: ret,
-		})
-		if err != nil {
-			logger.Error(err, "could not insert return uprobe. Skipping")
-			return
-		}
-		h.returnProbs = append(h.returnProbs, retProbe)
-	}
-}
-
 // Run runs the events processing loop.
 func (h *Instrumentor) Run(eventsChan chan<- *events.Event) {
-	logger := log.Logger.WithName("net/http-instrumentor")
+	logger := log.Logger.WithName("database/sql/sql-instrumentor")
 	var event Event
 	for {
 		record, err := h.eventsReader.Read()
@@ -199,8 +169,7 @@ func (h *Instrumentor) Run(eventsChan chan<- *events.Event) {
 }
 
 func (h *Instrumentor) convertEvent(e *Event) *events.Event {
-	method := unix.ByteSliceToString(e.Method[:])
-	path := unix.ByteSliceToString(e.Path[:])
+	query := unix.ByteSliceToString(e.Query[:])
 
 	sc := trace.NewSpanContext(trace.SpanContextConfig{
 		TraceID:    e.SpanContext.TraceID,
@@ -222,25 +191,22 @@ func (h *Instrumentor) convertEvent(e *Event) *events.Event {
 	}
 
 	return &events.Event{
-		Library: h.LibraryName(),
-		// Do not include the high-cardinality path here (there is no
-		// templatized path manifest to reference).
-		Name:              method,
-		Kind:              trace.SpanKindServer,
-		StartTime:         int64(e.StartTime),
-		EndTime:           int64(e.EndTime),
-		SpanContext:       &sc,
-		ParentSpanContext: pscPtr,
+		Library:     h.LibraryName(),
+		Name:        "DB",
+		Kind:        trace.SpanKindClient,
+		StartTime:   int64(e.StartTime),
+		EndTime:     int64(e.EndTime),
+		SpanContext: &sc,
 		Attributes: []attribute.KeyValue{
-			semconv.HTTPMethodKey.String(method),
-			semconv.HTTPTargetKey.String(path),
+			semconv.DBStatementKey.String(query),
 		},
+		ParentSpanContext: pscPtr,
 	}
 }
 
 // Close stops the Instrumentor.
 func (h *Instrumentor) Close() {
-	log.Logger.V(0).Info("closing net/http instrumentor")
+	log.Logger.V(0).Info("closing database/sql/sql instrumentor")
 	if h.eventsReader != nil {
 		h.eventsReader.Close()
 	}
@@ -256,4 +222,17 @@ func (h *Instrumentor) Close() {
 	if h.bpfObjects != nil {
 		h.bpfObjects.Close()
 	}
+}
+
+// shouldIncludeDBStatement returns if the user has configured SQL queries to be included.
+func shouldIncludeDBStatement() bool {
+	val := os.Getenv(IncludeDBStatementEnvVar)
+	if val != "" {
+		boolVal, err := strconv.ParseBool(val)
+		if err == nil {
+			return boolVal
+		}
+	}
+
+	return false
 }